pyproject.toml
[project]
name = "momoney"
version = "0.1.0"
description = "Self-hosted personal finance tracker"
requires-python = ">=3.11"
dependencies = [
    "watchdog>=3.0.0",
    "anthropic>=0.18.0",
    "gspread>=5.12.0",
    "google-auth>=2.25.0",
    "google-api-python-client>=2.100.0",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]

[project.scripts]
momoney = "src.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
